Fix GH-2201: Drop imagesx() return false description
As of PHP 8.0.0, GD images are objects, while previously they have been
resources.  If incompatible objects are passed to respective GD
functions, that throws a `TypeError` which does not have to be
documented explicitly.  If invalid resources have been passed, that
emitted `E_WARNING` and the functions returned `false`.  In a perfect
world, we would document that for each function, but to be brief, we
just augment the respective language-snippet to express that valid `gd`
resources have been expected.  Readers should be able to figure out the
detailed behavior (and would want to avoid passing invalid resources
anyway).